+++ src/test/ExSln2/c99/gen/hal_LedArray.h
@@ -1,9 +1,9 @@
 // finlang generated file for c# hal.LedArray type
 #pragma once
 
+#include "hal_Led.h"
 #include <stdint.h>
 #include "hal_LedArray_LedInfo.h"
-#include "hal_Led.h"