+++ src/test/ExSln2/LedBlinker/app/Main.cs
@@ -42,6 +42,7 @@ public class Main : FinObj
         b = (a + my_i32).narrow_to_u8();
         b = (u8)my_i32;
         b = u8.narrow_from(my_i32);
+        FinC.ignore_unused(b);
 
         return _toggle_at_ms.u32_ + 20; //u32_ is self-declaration
     }