+++ src/test/ExSln2/LedBlinker/hal/ImplicitInterfaceConversions.cs
@@ -1,4 +1,5 @@
 ﻿using finlang;
+using static finlang.FinC; // for ignore_unused()
 
 namespace hal;
 
@@ -15,9 +16,13 @@ public class ImplicitInterfaceConversions : FinObj
 
         take_dig_in(dig_in); // shouldn't have a conversion
         IDigIn dig_in2 = dig_in; // shouldn't have a conversion
+
+        FinC.ignore_unused(dig_out);
+        ignore_unused(dig_in2);
     }
 
     public static void take_dig_in(IDigIn dig_in)
     {
+         FinC.ignore_unused(dig_in);
     }
 }