+++ src/test/ExSln2/c99/gen/app_Main.c
@@ -41,6 +41,7 @@ uint32_t app_Main_transpilation_test_stuff(app_Main * self)
     b = (uint8_t)(a + my_i32);
     b = (uint8_t)my_i32;
     b = (uint8_t)my_i32;
+    (void)(b);
 
     return self->_toggle_at_ms + 20; //u32_ is self-declaration
 }